refactor: extract App Ecosystem and Top Hosted Apps to shared components
Moves AppEcosystemPanel and TopHostedAppsPanel out of home/HomeOverview
into standalone components/AppEcosystemBreakdown and
components/TopHostedApps so they can be reused by the upcoming
Analytics page (AppsTab). Behavior-preserving: HomeOverview's two
render call sites now use the new components with the same gstore
prop, and the moved CSS lives in each component's own index.scss.
