+++ client/src/home/HomeOverview/index.scss
@@ -81,8 +81,6 @@
   // Per-panel accent strip on left edge
   &--stats      { border-left: 2px solid #2686d0; }
   &--resources  { border-left: 2px solid #6366f1; }
-  &--ecosystem  { border-left: 2px solid #f59e0b; }
-  &--top-apps   { border-left: 2px solid #10b981; }
   &--expiring   { border-left: 2px solid #ef4444; }
   &--deployed   { border-left: 2px solid #22c55e; }
   &--geo        { border-left: 2px solid #8b5cf6; }
@@ -136,8 +134,6 @@
 // Per-panel accent gradient tint behind header (reduced opacity for subtlety)
 .hov-panel--stats     .hov-header { background: linear-gradient(90deg, rgba(38, 134, 208, 0.04) 0%, transparent 60%); }
 .hov-panel--resources .hov-header { background: linear-gradient(90deg, rgba(99, 102, 241, 0.04) 0%, transparent 60%); }
-.hov-panel--ecosystem .hov-header { background: linear-gradient(90deg, rgba(245, 158, 11, 0.04) 0%, transparent 60%); }
-.hov-panel--top-apps  .hov-header { background: linear-gradient(90deg, rgba(16, 185, 129, 0.04) 0%, transparent 60%); }
 .hov-panel--expiring  .hov-header { background: linear-gradient(90deg, rgba(239, 68, 68, 0.04) 0%, transparent 60%); }
 .hov-panel--deployed  .hov-header { background: linear-gradient(90deg, rgba(34, 197, 94, 0.04) 0%, transparent 60%); }
 .hov-panel--geo       .hov-header { background: linear-gradient(90deg, rgba(139, 92, 246, 0.04) 0%, transparent 60%); }
@@ -310,203 +306,15 @@
   font-variant-numeric: tabular-nums;
 }
 
-// ── App Ecosystem panel ────────────────────────────────────────────────────────
-
-.hov-eco-list {
-  display: flex;
-  flex-direction: column;
-  gap: 6px;
-}
-
-.hov-eco-row {
-  display: grid;
-  align-items: center;
-  gap: 5px;
-  grid-template-columns: 14px 80px 1fr 40px 30px 14px;
-  cursor: pointer;
-  border-radius: 4px;
-  transition: background var(--transition-fast);
-
-  &:hover,
-  &:focus-visible {
-    background: var(--surface-inset);
-    outline: none;
-  }
-
-  &--expanded {
-    background: var(--surface-inset);
-  }
-
-  // Discrete weight steps — a big category reads bigger, a tiny one reads
-  // smaller, instead of every row sharing identical type/icon size.
-  &--lg .hov-eco-label { font-size: 0.86rem; font-weight: 600; }
-  &--lg .hov-eco-icon svg { width: 13px; height: 13px; }
-  &--lg .hov-eco-count { font-size: 0.8rem; }
-
-  &--sm .hov-eco-label { font-size: 0.72rem; opacity: 0.85; }
-  &--sm .hov-eco-icon svg { width: 10px; height: 10px; }
-  &--sm .hov-eco-count { font-size: 0.7rem; opacity: 0.85; }
-}
-
-.hov-eco-icon {
-  display: flex;
-  align-items: center;
-  flex-shrink: 0;
-}
-
-.hov-eco-label {
-  font-size: 0.78rem;
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  cursor: help;
-  border-bottom: 1px dashed rgba(128, 128, 128, 0.35);
-}
-
-.hov-eco-chevron {
-  color: var(--text-tertiary);
-  transition: transform var(--transition-fast);
-  flex-shrink: 0;
-
-  &--open {
-    transform: rotate(180deg);
-  }
-}
-
-.hov-eco-expand {
-  padding: 6px 8px 10px 19px;
-  animation: hov-eco-expand-in 0.15s ease-out;
-
-  // CategoryTooltip is normally only ever seen inside Blueprint's dark
-  // Tooltip2 popover, where its fixed light border/opacity read fine. Placed
-  // inline on the panel's own themed surface instead, give it theme-aware
-  // colors so it doesn't nearly vanish in light mode.
-  .cat-tooltip {
-    max-width: none;
-    color: var(--text-secondary);
-  }
-
-  .cat-tooltip__list {
-    border-top-color: var(--border-primary);
-  }
-}
-
-@keyframes hov-eco-expand-in {
-  from { opacity: 0; transform: translateY(-4px); }
-  to { opacity: 1; transform: translateY(0); }
-}
-
-.hov-eco-bar-wrap {
-  height: 5px;
-  border-radius: 3px;
-  background: var(--surface-inset);
-  overflow: hidden;
-}
-
-.hov-eco-bar-fill {
-  height: 100%;
-  border-radius: 3px;
-  transition: width 0.5s cubic-bezier(0.4, 0, 0.2, 1);
-  opacity: 0.85;
-}
-
-.hov-eco-count {
-  font-size: 0.75rem;
-  font-weight: 600;
-  font-variant-numeric: tabular-nums;
-  text-align: right;
-}
-
-.hov-eco-pct {
-  font-size: 0.68rem;
-  color: var(--text-tertiary);
-  text-align: right;
-}
-
-// ── Ranked list (top hosted apps) ─────────────────────────────────────────────
-
-.hov-ranked-list {
-  display: flex;
-  flex-direction: column;
-  gap: 5px;
-}
-
-.hov-ranked-row {
-  display: grid;
-  align-items: center;
-  gap: 6px;
-  grid-template-columns: 20px 1fr 60px 36px;
-  padding: 4px 6px;
-  border-radius: var(--radius-sm);
-  border-bottom: 1px solid rgba(128, 128, 128, 0.07);
-  transition: background var(--transition-fast);
-
-  &:hover {
-    background: var(--surface-inset);
-  }
-
-  &:last-child {
-    border-bottom: none;
-  }
-}
-
-.hov-rank {
-  font-size: 0.65rem;
-  color: var(--text-tertiary);
-  font-weight: 700;
-  font-variant-numeric: tabular-nums;
-  width: 20px;
-  height: 20px;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 50%;
-
-  &--gold {
-    color: #d97706;
-    background: rgba(217, 119, 6, 0.12);
-  }
-  &--silver {
-    color: #94a3b8;
-    background: rgba(148, 163, 184, 0.12);
-  }
-  &--bronze {
-    color: #b45309;
-    background: rgba(180, 83, 9, 0.12);
-  }
-}
-
-.hov-ranked-name {
-  font-size: 0.78rem;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  min-width: 0;
-}
-
-.hov-ranked-bar-wrap {
-  height: 4px;
-  border-radius: 2px;
-  background: var(--surface-inset);
-  overflow: hidden;
-}
-
-.hov-ranked-bar-fill {
-  height: 100%;
-  border-radius: 2px;
-  background: linear-gradient(90deg, #2686d0, #5eb8ff);
-  transition: width 0.5s cubic-bezier(0.4, 0, 0.2, 1);
-}
-
 // ── Bottom row: fixed 10-row height panels ────────────────────────────────────
+// (.hov-panel--top-apps's slice of this rule now lives in
+// components/TopHostedApps/index.scss, alongside its own .hov-ranked-list.)
 
-.hov-panel--top-apps,
 .hov-panel--expiring,
 .hov-panel--deployed {
   display: flex;
   flex-direction: column;
 
-  .hov-ranked-list,
   .hov-list {
     height: 380px;
     overflow-y: auto;
@@ -710,20 +518,6 @@
       color: #4ade80;
     }
   }
-}
-
-.hov-eco-stale {
-  margin-left: auto;
-  margin-right: 8px;
-  font-size: 0.6rem;
-  font-weight: 600;
-  letter-spacing: 0.06em;
-  text-transform: uppercase;
-  color: #f97316;
-  border: 1px solid rgba(249, 115, 22, 0.45);
-  border-radius: 3px;
-  padding: 1px 5px;
-  cursor: help;
 }
 
 .hov-empty {
